docs :📚 테스트코드 폴더로 구분
- TravelTests, CountryTests로 파일을 구분
- CountryProviderStub은 불필요하여 제거
- CountryStub 클래스 정의부분을 별도의 파일로 관리
