--- BoostPocket/BoostPocketTests/CountryProviderStub.swift
@@ -1,12 +1,0 @@
-//
-//  CountryProviderStub.swift
-//  BoostPocketTests
-//
-//  Created by sihyung you on 2020/11/23.
-//  Copyright © 2020 BoostPocket. All rights reserved.
-//
-
-import Foundation
-@testable import BoostPocket
-
-class CountryProviderStub: CountryProvider { }
